SGC/requirements.txt
amqp==5.1.1
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.1
django-cors-headers==3.11.0
django-dbbackup==4.0.0b0
djangorestframework==3.13.1
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
html5lib==1.1
kombu==5.2.4
prompt-toolkit==3.0.31
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
XlsxWriter==3.0.3
zope.event==4.5.0
zope.interface==5.4.0
